+++ src/main/java/org/example/service/ProfileService.java
@@ -25,7 +25,7 @@ public class ProfileService {
             return toDTO(save);
     }
 
-    public ProfileEntity get(Integer id) {
+    public ProfileEntity get(Long id) {
         return profileRepository.findById(id).orElseThrow(() -> {
             logger.error("Profile not found id = {}", id);
             throw new AppBadException("Profile not found");
@@ -55,7 +55,7 @@ public class ProfileService {
     }
 
     public Boolean changePassword(ProfileChangePasswordDTO dto) {
-        Integer profileId = SecurityUtil.getProfileId();
+        Long profileId = SecurityUtil.getProfileId();
         ProfileEntity entity= get(profileId);
         if (!entity.getPassword().equals(MD5Util.getMD5(entity.getPassword()))){
            throw new AppBadException("Old password wrong");
